Migrate queries to builder classes
